backend/requirements.txt
accelerate==1.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
amqp==5.3.1
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.1
CacheControl==0.14.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
channels==4.1.0
channels_redis==4.2.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==44.0.2
daphne==4.1.2
Django==5.1.2
django-cors-headers==4.5.0
django-redis==5.4.0
django-webpush==0.3.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filelock==3.18.0
firebase-admin==6.7.0
frozenlist==1.5.0
fsspec==2025.3.2
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.1
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.71.0
h2==4.2.0
hf-xet==1.0.2
hpack==4.1.0
http_ece==1.2.1
httplib2==0.22.0
huggingface-hub==0.30.1
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
Jinja2==3.1.6
kombu==5.5.0
MarkupSafe==3.0.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.2.0
networkx==3.4.2
numpy==2.2.4
packaging==24.2
pillow==11.0.0
priority==1.3.0
prompt_toolkit==3.0.50
propcache==0.3.0
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
py-vapid==1.9.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==25.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pywebpush==2.0.3
PyYAML==6.0.2
redis==5.2.1
regex==2024.11.6
requests==2.32.3
rsa==4.9
safetensors==0.5.3
service-identity==24.2.0
setuptools==78.1.0
six==1.17.0
sqlparse==0.5.1
sympy==1.13.1
tokenizers==0.21.1
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.51.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.13.1
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
yarl==1.18.3
zope.interface==7.2
psycopg2-binary
dj-database-url   
gdown